+++ src/IncrementalChecker.ts
@@ -1,5 +1,5 @@
 import * as fs from 'fs';
-import endsWith = require('lodash.endswith');
+import endsWith = require('lodash/endsWith');
 import * as path from 'path';
 import * as ts from 'typescript';
 import { Configuration, Linter } from 'tslint'; // Imported for types alone; actual requires take place in methods below
@@ -22,6 +22,7 @@ interface ConfigurationFile extends Configuration.IConfigurationFile {
 
 export class IncrementalChecker {
   programConfigFile: string;
+  compilerOptions: object;
   linterConfigFile: string | false;
   watchPaths: string[];
   workNumber: number;
@@ -41,6 +42,7 @@ export class IncrementalChecker {
 
   constructor(
     programConfigFile: string,
+    compilerOptions: object,
     linterConfigFile: string | false,
     watchPaths: string[],
     workNumber: number,
@@ -49,6 +51,7 @@ export class IncrementalChecker {
     vue: boolean
   ) {
     this.programConfigFile = programConfigFile;
+    this.compilerOptions = compilerOptions;
     this.linterConfigFile = linterConfigFile;
     this.watchPaths = watchPaths;
     this.workNumber = workNumber || 0;
@@ -61,26 +64,37 @@ export class IncrementalChecker {
 
     // it's shared between compilations
     this.files = new FilesRegister(() => ({
-        // data shape
-        source: undefined,
-        linted: false,
-        lints: []
+      // data shape
+      source: undefined,
+      linted: false,
+      lints: []
     }));
   }
 
-  static loadProgramConfig(configFile: string) {
-    return ts.parseJsonConfigFileContent(
-      // Regardless of the setting in the tsconfig.json we want isolatedModules to be false
-      Object.assign(ts.readConfigFile(configFile, ts.sys.readFile).config, { isolatedModules: false }),
+  static loadProgramConfig(configFile: string, compilerOptions: object) {
+    const tsconfig = ts.readConfigFile(configFile, ts.sys.readFile).config;
+
+    tsconfig.compilerOptions = tsconfig.compilerOptions || {};
+    tsconfig.compilerOptions = {
+      ...tsconfig.compilerOptions,
+      ...compilerOptions
+    };
+
+    const parsed = ts.parseJsonConfigFileContent(
+      tsconfig,
       ts.sys,
       path.dirname(configFile)
     );
+
+    return parsed;
   }
 
   static loadLinterConfig(configFile: string): ConfigurationFile {
     const tslint = require('tslint');
 
-    return tslint.Configuration.loadConfigurationFromPath(configFile) as ConfigurationFile;
+    return tslint.Configuration.loadConfigurationFromPath(
+      configFile
+    ) as ConfigurationFile;
   }
 
   static createProgram(
@@ -107,7 +121,7 @@ export class IncrementalChecker {
 
       // get source file only if there is no source in files register
       if (!files.has(filePath) || !files.getData(filePath).source) {
-        files.mutateData(filePath, (data) => {
+        files.mutateData(filePath, data => {
           data.source = realGetSourceFile(filePath, languageVersion, onError);
         });
       }
@@ -129,13 +143,21 @@ export class IncrementalChecker {
     return new tslint.Linter({ fix: false }, program);
   }
 
-  static isFileExcluded(filePath: string, linterExclusions: minimatch.IMinimatch[]): boolean {
-    return endsWith(filePath, '.d.ts') || linterExclusions.some(matcher => matcher.match(filePath));
+  static isFileExcluded(
+    filePath: string,
+    linterExclusions: minimatch.IMinimatch[]
+  ): boolean {
+    return (
+      endsWith(filePath, '.d.ts') ||
+      linterExclusions.some(matcher => matcher.match(filePath))
+    );
   }
 
   nextIteration() {
     if (!this.watcher) {
-      const watchExtensions = this.vue ? ['.ts', '.tsx', '.vue'] : ['.ts', '.tsx'];
+      const watchExtensions = this.vue
+        ? ['.ts', '.tsx', '.vue']
+        : ['.ts', '.tsx'];
       this.watcher = new FilesWatcher(this.watchPaths, watchExtensions);
 
       // connect watcher with register
@@ -150,13 +172,20 @@ export class IncrementalChecker {
     }
 
     if (!this.linterConfig && this.linterConfigFile) {
-      this.linterConfig = IncrementalChecker.loadLinterConfig(this.linterConfigFile);
+      this.linterConfig = IncrementalChecker.loadLinterConfig(
+        this.linterConfigFile
+      );
 
-      if (this.linterConfig.linterOptions && this.linterConfig.linterOptions.exclude) {
+      if (
+        this.linterConfig.linterOptions &&
+        this.linterConfig.linterOptions.exclude
+      ) {
         // Pre-build minimatch patterns to avoid additional overhead later on.
         // Note: Resolving the path is required to properly match against the full file paths,
         // and also deals with potential cross-platform problems regarding path separators.
-        this.linterExclusions = this.linterConfig.linterOptions.exclude.map(pattern => new minimatch.Minimatch(path.resolve(pattern)));
+        this.linterExclusions = this.linterConfig.linterOptions.exclude.map(
+          pattern => new minimatch.Minimatch(path.resolve(pattern))
+        );
       }
     }
 
@@ -168,7 +197,12 @@ export class IncrementalChecker {
   }
 
   loadVueProgram() {
-    this.programConfig = this.programConfig || VueProgram.loadProgramConfig(this.programConfigFile);
+    this.programConfig =
+      this.programConfig ||
+      VueProgram.loadProgramConfig(
+        this.programConfigFile,
+        this.compilerOptions
+      );
 
     return VueProgram.createProgram(
       this.programConfig,
@@ -180,9 +214,19 @@ export class IncrementalChecker {
   }
 
   loadDefaultProgram() {
-    this.programConfig = this.programConfig || IncrementalChecker.loadProgramConfig(this.programConfigFile);
+    this.programConfig =
+      this.programConfig ||
+      IncrementalChecker.loadProgramConfig(
+        this.programConfigFile,
+        this.compilerOptions
+      );
 
-    return IncrementalChecker.createProgram(this.programConfig, this.files, this.watcher, this.program);
+    return IncrementalChecker.createProgram(
+      this.programConfig,
+      this.files,
+      this.watcher,
+      this.program
+    );
   }
 
   hasLinter() {
@@ -195,7 +239,11 @@ export class IncrementalChecker {
     const filesToCheck = this.program.getSourceFiles();
 
     // calculate subset of work to do
-    const workSet = new WorkSet(filesToCheck, this.workNumber, this.workDivision);
+    const workSet = new WorkSet(
+      filesToCheck,
+      this.workNumber,
+      this.workDivision
+    );
 
     // check given work set
     workSet.forEach(sourceFile => {
@@ -203,10 +251,18 @@ export class IncrementalChecker {
         cancellationToken.throwIfCancellationRequested();
       }
 
-      const diagnosticsToRegister: ReadonlyArray<ts.Diagnostic> = this.checkSyntacticErrors
+      const diagnosticsToRegister: ReadonlyArray<ts.Diagnostic> = this
+        .checkSyntacticErrors
         ? []
-          .concat(this.program.getSemanticDiagnostics(sourceFile, cancellationToken))
-          .concat(this.program.getSyntacticDiagnostics(sourceFile, cancellationToken))
+            .concat(
+              this.program.getSemanticDiagnostics(sourceFile, cancellationToken)
+            )
+            .concat(
+              this.program.getSyntacticDiagnostics(
+                sourceFile,
+                cancellationToken
+              )
+            )
         : this.program.getSemanticDiagnostics(sourceFile, cancellationToken);
 
       diagnostics.push.apply(diagnostics, diagnosticsToRegister);
@@ -224,12 +280,20 @@ export class IncrementalChecker {
     }
 
     // select files to lint
-    const filesToLint = this.files.keys().filter(filePath =>
-      !this.files.getData(filePath).linted && !IncrementalChecker.isFileExcluded(filePath, this.linterExclusions)
-    );
+    const filesToLint = this.files
+      .keys()
+      .filter(
+        filePath =>
+          !this.files.getData(filePath).linted &&
+          !IncrementalChecker.isFileExcluded(filePath, this.linterExclusions)
+      );
 
     // calculate subset of work to do
-    const workSet = new WorkSet(filesToLint, this.workNumber, this.workDivision);
+    const workSet = new WorkSet(
+      filesToLint,
+      this.workNumber,
+      this.workDivision
+    );
 
     // lint given work set
     workSet.forEach(fileName => {
@@ -239,11 +303,11 @@ export class IncrementalChecker {
         this.linter.lint(fileName, undefined, this.linterConfig);
       } catch (e) {
         if (
-            fs.existsSync(fileName) &&
-            // check the error type due to file system lag
-            !(e instanceof Error) &&
-            !(e.constructor.name === 'FatalError') &&
-            !(e.message && e.message.trim().startsWith("Invalid source file"))
+          fs.existsSync(fileName) &&
+          // check the error type due to file system lag
+          !(e instanceof Error) &&
+          !(e.constructor.name === 'FatalError') &&
+          !(e.message && e.message.trim().startsWith('Invalid source file'))
         ) {
           // it's not because file doesn't exist - throw error
           throw e;
@@ -269,9 +333,13 @@ export class IncrementalChecker {
     });
 
     // get all lints
-    const lints = this.files.keys().reduce((innerLints, filePath) =>
-      innerLints.concat(this.files.getData(filePath).lints),
-      []);
+    const lints = this.files
+      .keys()
+      .reduce(
+        (innerLints, filePath) =>
+          innerLints.concat(this.files.getData(filePath).lints),
+        []
+      );
 
     // normalize and deduplicate lints
     return NormalizedMessage.deduplicate(